Fixed SOLID violation in TestService
