+++ src/main/java/com/testmaster/model/TestResult.java
@@ -16,4 +16,22 @@ public class TestResult {
     this.userTest = userTest;
     numberOfQuestions = userTest.getQuestions().size();
   }
+
+  public void submitAnswer(Question currentQuestion, int answer, TestResult testResult) {
+    answer--;
+    if (currentQuestion != null && answer >= 0
+        && answer < currentQuestion.getOptions().size()) {
+      if (checkAnswer(currentQuestion, answer)) {
+        testResult.setNumberOfCorrectAnswer(testResult.getNumberOfCorrectAnswer() + 1);
+      }
+    }
+  }
+
+  private boolean checkAnswer(Question currentQuestion, int answer) {
+    boolean correct = false;
+    if (currentQuestion.getOptions().get(answer).isCorrect()) {
+      correct =  true;
+    }
+    return correct;
+  }
 }
