+++ src/main/java/com/testmaster/service/TestService.java
@@ -1,7 +1,5 @@
 package com.testmaster.service;
 
-import com.testmaster.model.Question;
-import com.testmaster.model.TestResult;
 import com.testmaster.model.UserTest;
 import java.util.List;
 
@@ -10,5 +8,4 @@ public interface TestService {
 
   UserTest getTestByName(String testName);
 
-  void submitAnswer(Question currentQuestion, int answer, TestResult testResult);
 }
